% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Title}
\usage{
calibrate(c14_age, c14_error, calibration, precision = 1e-05)
}
\arguments{
\item{calibration}{an object of class \code{Calibration} built using \code{read_14c()}}

\item{precision}{a numeric scalar with the desired precision, meaning
the value at which smaller densities are set to zero (default 1e-5)}

\item{measured_age}{a numeric vector of uncalibrated radiocarbon ages}

\item{measured_error}{a numeric vector of uncalibrated radiocarbon errors}
}
\value{
a list of probability density distributions, with length equal to
\code{length(age)}
}
\description{
Title
}
