% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibration}
\alias{calibration}
\alias{as_calibration}
\alias{as_calibration.tbl}
\alias{as_calibration.data.frame}
\alias{is_calibration}
\alias{print.Calibration}
\title{Radiocarbon calibration curve}
\usage{
calibration(cal_age, est_age, est_error, calibration_name)

as_calibration(
  x,
  calendar_age,
  calibration_age,
  calibration_error,
  cal_name = "<unknown>"
)

\method{as_calibration}{tbl}(x, ...)

\method{as_calibration}{data.frame}(x, ...)

is_calibration(x)

\method{print}{Calibration}(x, ...)
}
\arguments{
\item{cal_age}{\itemize{
\item for \code{\link[=calibration]{calibration()}} a numeric vector of calendar ages
\item for \code{\link[=as_calibration]{as_calibration()}} a character scalar, the name of the column in \link{x}
with calendar ages
}}

\item{est_age}{\itemize{
\item for \code{\link[=calibration]{calibration()}} a numeric vector of estimated ages
\item for \code{\link[=as_calibration]{as_calibration()}} a character scalar, the name of the column in \link{x}
with estimated ages
}}

\item{est_error}{\itemize{
\item for \code{\link[=calibration]{calibration()}} a numeric vector of estimated errors
\item for \code{\link[=as_calibration]{as_calibration()}} a character scalar, the name of the column in \link{x}
with estimated errors
}}

\item{calibration_name}{a character scalar, the name of the radiocarbon
calibration curve being used}

\item{x}{\itemize{
\item for \code{\link[=is_calibration]{is_calibration()}} and \code{\link[=print]{print()}} a \link{Calibration} table
\item for \code{\link[=as_calibration]{as_calibration()}} a \link{tbl} or \link{data.frame} to convert to \link{Calibration}
}}

\item{n}{a numeric scalar, the number of rows in \link{data.frame} to print}
}
\value{
A \link{Calibration} table that inherits from \link{tibble} and \link{data.frame}. Has three
columns: cal_age, est_age, est_error.
}
\description{
Build a table of class \link{Calibration}. Additional methods for converting to
class, checking for class, and printing.
}
