% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate Calibration curve}
\usage{
interpolate(x, xout = NULL)
}
\arguments{
\item{x}{a \link{Calibration} table}

\item{xout}{years in which to estimate true ages. If NULL (default), then all
years in [seq(max(x[\link{"cal_age"}]), min(x[\link{"cal_age"}]), by = -1)] are used.}
}
\value{
A \link{Calibration} table that inherits from \link{tibble} and \link{data.frame}. Has three
columns: cal_age, est_age, est_error.
}
\description{
Radiocarbon calibration curves usually only estimate true ages at 10-year
(or greater) intervals. This function expands the curve to include estimates
in all intervening years using linear interpolation. Implemented in Rust for
additional speed.
}
