% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{rc_calibrate}
\alias{rc_calibrate}
\title{Calibrate radiocarbon measurements}
\usage{
rc_calibrate(
  c14_age,
  c14_error,
  calibration_tbl,
  precision = 1e-05,
  normalize = TRUE
)
}
\arguments{
\item{c14_age}{a numeric vector of uncalibrated radiocarbon ages}

\item{c14_error}{a numeric vector of uncalibrated radiocarbon errors}

\item{calibration_tbl}{a \code{Calibration} table}

\item{precision}{a numeric scalar with the desired precision, meaning
the value at which smaller densities are set to zero (default 1e-5)}

\item{normalize}{a logical scalar, if TRUE (default) then each probability
distribution is normalized.}
}
\value{
a \link{CalGrid}, strictly an integer vector of row IDs with an
attribute containing an external pointer to where the sparse matrix is held
in memory.
}
\description{
Calibrate radiocarbon measurements
}
