% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{rc_collect}
\alias{rc_collect}
\title{Move data into local R object}
\usage{
rc_collect(x)
}
\arguments{
\item{x}{a \code{CalGrid}}
}
\description{
Currently, returns a dense matrix, but will hopefully convert this to a
sparse matrix at some point.
}
