% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.R
\name{rc_spd}
\alias{rc_spd}
\title{Sum calibrated radiocarbon probability distributions}
\usage{
rc_spd(x, normalize = TRUE)
}
\arguments{
\item{x}{a \code{CalGrid} vector}

\item{normalize}{a logical scalar, if TRUE (default) then each probability
distribution is normalized.}
}
\value{
a \code{matrix} with two columns \code{ybp} and \code{prob_dens}
}
\description{
Sum calibrated radiocarbon probability distributions
}
