% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Title}
\usage{
calibrate(
  age,
  error,
  time_range = c(55000, 0),
  calibration = "intcal20",
  precision = 1e-05
)
}
\arguments{
\item{age}{a numeric vector of uncalibrated radiocarbon ages}

\item{error}{a numeric vector of uncalibrated radiocarbon errors}

\item{time_range}{a vector of length two with start and end dates
(note that start > end)}

\item{calibration}{a scalar character vector with the name of the calibration
curve}

\item{precision}{a scalar numeric vector with the desired precision, meaning
the value at which smaller densities are set to zero}
}
\value{
a list of probability density distributions, with length equal to
\code{length(age)}
}
\description{
Title
}
